feat: 認証、認可のサンプルを実装

+++ domain/src/main/kotlin/com/ikkun2501/bookmanagement/domain/Author.kt
@@ -8,7 +8,7 @@ package com.ikkun2501.bookmanagement.domain
  * @property description 説明
  */
 data class Author(
-    val authorId: Long,
+    val authorId: Int,
     val authorName: String,
     val description: String
 )
